Separate unit_tests from e2e tests.
