+++ package.json
@@ -152,7 +152,7 @@
     "watch": "watch 'clear && npm run lint' lib admin server test",
     "pretest-cov": "npm run clean && npm run lint",
     "current": "updtr",
-    "test-cov": "istanbul cover ./node_modules/mocha/bin/_mocha",
+    "test-cov": "istanbul cover -x 'test/e2e/**' ./node_modules/mocha/bin/_mocha",
     "posttest-cov": "if [ -n \"$CODECLIMATE_REPO_TOKEN\" ]; then codeclimate < coverage/lcov.info; fi",
     "clean": "rimraf ./coverage"
   },
