+++ test/e2e/nightwatch.json
@@ -1,5 +1,5 @@
 {
-  "src_folders" : ["test/e2e/tests"],
+  "src_folders" : ["test/e2e/test_groups"],
   "output_folder" : "reports",
   "custom_commands_path" : "",
   "custom_assertions_path" : "",
