+++ test/unit_tests/lib/path_test.js
@@ -1,5 +1,5 @@
 var demand = require('must');
-var Path = require("../../lib/path");
+var Path = require("../../../lib/path");
 
 describe("Path", function() {
 	describe("new", function() {
